+++ gotham/src/service/mod.rs
@@ -10,7 +10,6 @@ use hyper;
 use hyper::server::Service;
 use hyper::{Request, Response};
 use futures::Future;
-use tokio::reactor::Handle;
 
 use handler::NewHandler;
 use state::{request_id, set_request_id, State};
@@ -27,21 +26,19 @@ where
     T: NewHandler + 'static,
 {
     t: Arc<T>,
-    handle: Handle,
 }
 
 impl<T> GothamService<T>
 where
     T: NewHandler + 'static,
 {
-    pub(crate) fn new(t: Arc<T>, handle: Handle) -> GothamService<T> {
-        GothamService { t, handle }
+    pub(crate) fn new(t: Arc<T>) -> GothamService<T> {
+        GothamService { t }
     }
 
     pub(crate) fn connect(&self, client_addr: SocketAddr) -> ConnectedGothamService<T> {
         ConnectedGothamService {
             t: self.t.clone(),
-            handle: self.handle.clone(),
             client_addr,
         }
     }
@@ -54,7 +51,6 @@ where
     T: NewHandler + 'static,
 {
     t: Arc<T>,
-    handle: Handle,
     client_addr: SocketAddr,
 }
 
@@ -74,7 +70,6 @@ where
 
         let (method, uri, version, headers, body) = req.deconstruct();
 
-        state.put(self.handle.clone());
         state.put(RequestPathSegments::new(uri.path()));
         state.put(method);
         state.put(uri);
@@ -98,7 +93,7 @@ mod tests {
     use super::*;
 
     use hyper::{Method, StatusCode};
-    use tokio::runtime::Runtime;
+    use tokio;
 
     use http::response::create_response;
     use router::builder::*;
@@ -111,8 +106,7 @@ mod tests {
 
     #[test]
     fn new_handler_closure() {
-        let mut core = Core::new().unwrap();
-        let service = GothamService::new(Arc::new(|| Ok(handler)), core.handle());
+        let service = GothamService::new(Arc::new(|| Ok(handler)));
 
         let req = Request::new(Method::Get, "http://localhost/".parse().unwrap());
         let f = service
@@ -129,7 +123,7 @@ mod tests {
         });
 
         let mut core = Core::new().unwrap();
-        let service = GothamService::new(Arc::new(router), core.handle());
+        let service = GothamService::new(Arc::new(router));
 
         let req = Request::new(Method::Get, "http://localhost/".parse().unwrap());
         let f = service
