+++ gotham/src/test/mod.rs
@@ -3,14 +3,13 @@
 //! See the `TestServer` type for example usage.
 
 use std::{cell, io, net, time};
-use std::cell::RefCell;
 use std::net::{IpAddr, SocketAddr, TcpListener, TcpStream};
 use std::ops::{Deref, DerefMut};
 use std::rc::Rc;
 use std::sync::Arc;
 
 use futures::{future, Future, Stream};
-use futures_timer::ext::Timeout;
+use futures_timer::Delay;
 use hyper::{self, Body, Method, Request, Response, Uri};
 use hyper::client::{self, Client};
 use hyper::error::UriError;
@@ -18,10 +17,9 @@ use hyper::header::ContentType;
 use hyper::server::{self, Http};
 use mime;
 use mio;
-use tokio::executor::Executor;
+use tokio;
 use tokio::executor::thread_pool::{Builder as ThreadPoolBuilder};
-use tokio::reactor::PollEvented;
-use tokio::runtime::Runtime;
+use tokio::reactor::PollEvented2;
 
 use handler::NewHandler;
 use service::GothamService;
@@ -68,7 +66,6 @@ struct TestServerData<NH = Router>
 where
     NH: NewHandler + 'static,
 {
-    executor: RefCell<Executor>,
     http: Http,
     timeout: u64,
     gotham_service: GothamService<NH>,
@@ -123,13 +120,10 @@ where
     pub fn with_timeout(new_handler: NH, timeout: u64) -> Result<TestServer<NH>, io::Error> {
         let pool = ThreadPoolBuilder::new().pool_size(1).build();
 
-        let handle = pool.handle();
-
         let data = TestServerData {
-            executor: RefCell::new(pool),
             http: server::Http::new(),
             timeout,
-            gotham_service: GothamService::new(Arc::new(new_handler), handle),
+            gotham_service: GothamService::new(Arc::new(new_handler)),
         };
 
         Ok(TestServer {
@@ -153,8 +147,6 @@ where
     }
 
     fn try_client_with_address(&self, client_addr: net::SocketAddr) -> io::Result<TestClient<NH>> {
-        let handle = self.data.core.borrow().handle();
-
         let (cs, ss) = {
             // We're creating a private TCP-based pipe here. Bind to an ephemeral port, connect to
             // it and then immediately discard the listener.
@@ -166,10 +158,10 @@ where
         };
 
         let cs = mio::net::TcpStream::from_stream(cs)?;
-        let cs = PollEvented::new(cs, &handle)?;
+        let cs = PollEvented2::new(cs);
 
         let ss = mio::net::TcpStream::from_stream(ss)?;
-        let ss = PollEvented::new(ss, &handle)?;
+        let ss = PollEvented2::new(ss);
 
         let service = self.data.gotham_service.connect(client_addr);
         let f = self.data
@@ -178,7 +170,7 @@ where
             .map(|_| ())
             .map_err(|_| ());
 
-        handle.spawn(f);
+        tokio::spawn(f);
 
         let client = Client::configure()
             .connector(TestConnect {
@@ -201,13 +193,9 @@ where
         F: Future<Error = hyper::Error>,
     {
         let timeout_duration = time::Duration::from_secs(self.data.timeout);
-        let timeout = Timeout::new(timeout_duration, &self.data.core.borrow().handle())
-            .map_err(|e| TestRequestError::IoError(e))?;
+        let timeout = Delay::new(timeout_duration);
 
-        let run_result = {
-            let mut core = self.data.core.borrow_mut();
-            core.run(f.select2(timeout))
-        };
+        let run_result = tokio::run(f.select2(timeout));
 
         match run_result {
             Ok(future::Either::A((item, _))) => Ok(item),
@@ -229,8 +217,7 @@ where
             let f: hyper::Body = response.body();
             let f = f.for_each(|chunk| future::ok(buf.extend(chunk.into_iter())));
 
-            let mut core = self.data.core.borrow_mut();
-            core.run(f)
+            tokio::run(f)
         };
 
         r.map(|_| buf)
@@ -441,13 +428,13 @@ impl TestResponse {
 /// `TestConnect` represents the connection between a test client and the `TestServer` instance
 /// that created it. This type should never be used directly.
 struct TestConnect {
-    stream: cell::RefCell<Option<PollEvented<mio::net::TcpStream>>>,
+    stream: cell::RefCell<Option<PollEvented2<mio::net::TcpStream>>>,
 }
 
 impl client::Service for TestConnect {
     type Request = hyper::Uri;
     type Error = io::Error;
-    type Response = PollEvented<mio::net::TcpStream>;
+    type Response = PollEvented2<mio::net::TcpStream>;
     type Future = future::FutureResult<Self::Response, Self::Error>;
 
     fn call(&self, _req: Self::Request) -> Self::Future {
